refactoring modalitem and headerlayout components width css modules
