+++ src/App.js
@@ -1,9 +1,9 @@
 
-import { ThemeProvider, createTheme } from '@mui/material';
 import React from 'react';
+import { ThemeProvider, createTheme } from '@mui/material';
 import { Routes, Route } from 'react-router-dom'
 import './App.css';
-import HeaderContainer from './components/cosmetics/headerLayout/headerLayout';
+import HeaderContainer from './components/cosmetics/headerLayout/HeaderLayout';
 import TypeLayout from './components/cosmetics/typeLayout';
 import HomePage from './components/home/homePage';
 
@@ -13,7 +13,6 @@ const darkTheme = createTheme({
     mode: 'dark',
     background: { paper: '#2f3136' },
     primary: { main: '#fce200' },
-    //secondary: { light: '#fce200' }
     text: { primary: '#fce200' }
   },
 
@@ -23,15 +22,12 @@ const App = () => {
   return (
     <ThemeProvider theme={darkTheme}>
       <div className="App">
-
-
         <Routes>
           <Route path='/' element={<HeaderContainer />}>
             <Route index element={<HomePage />} />
-            <Route path='cosmetics/*' element={< TypeLayout />} />
+            <Route path='cosmetics' element={< TypeLayout />} />
           </Route>
         </Routes>
-
       </div>
     </ThemeProvider>
   );
