+++ src/components/cosmetics/modalItem/modalItem.module.css
@@ -4,7 +4,7 @@
   justify-content: center;
 }
 .wrapper {
-  box-shadow: 24;
+  /*box-shadow: 5;*/
   padding: 4;
   display: flex;
   align-items: center;
@@ -30,7 +30,7 @@
   text-align: center;
 }
 .closeBtn {
-  position: absolute;
+  position: absolute !important;
   top: 1rem;
   right: 1rem;
 }
